requirements.txt
pyyaml
nbformat
openai
python-dotenv
